mysite/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.6
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
django-filter==25.2
django-modeltranslation==0.20.2
django-phonenumber-field==8.4.0
django-rest-swagger==2.2.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==26.0
phonenumbers==9.0.26
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
simplejson==3.20.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
